YJDH-93: move auth components and common code to shared project
