+++ frontend/kesaseteli/employer/src/pages/application.tsx
@@ -1,11 +1,11 @@
 import CompanyInfo from 'kesaseteli/employer/components/companyInfo/CompanyInfo';
 import EmployerForm from 'kesaseteli/employer/components/employerForm/EmployerForm';
-import withAuth from 'kesaseteli/employer/hocs/withAuth';
 import { GetStaticProps, NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useTranslation } from 'next-i18next';
 import * as React from 'react';
 import FormSection from 'shared/components/forms/section/FormSection';
+import withAuth from 'shared/components/hocs/withAuth';
 import Layout from 'shared/components/Layout';
 import getServerSideTranslations from 'shared/i18n/get-server-side-translations';
 import { DEFAULT_LANGUAGE } from 'shared/i18n/i18n';
