+++ frontend/shared/src/auth/AuthContext.tsx
