+++ frontend/shared/src/components/__tests__/Layout.test.tsx
@@ -4,7 +4,9 @@ import Layout from 'shared/components/Layout';
 import { render } from 'test-utils';
 
 test('test for accessibility violations', async () => {
-  const { container } = render(<Layout headingText="Hello"> ... world!</Layout>);
+  const { container } = render(
+    <Layout headingText="Hello"> ... world!</Layout>
+  );
   const results = await axe(container);
   expect(results).toHaveNoViolations();
 });
