+++ frontend/shared/src/components/hocs/withAuth.tsx
